Start separating services from docker
Part of this change is that I want to limit the dependence on Docker,
especially for the api and frontend services. This is because I am
having trouble with getting docker to do hot reloading the way I'd like.

So instead of using Docker for everything, I'm just going to use it for
the parts I don't really want to set up myself: PostgreSQL, Nginx, and
Redis -- basically any pre-configured services that we're not writing
ourselves.

There are a few benefits to this approach. The most important one, in my
opinion, is that it's going to be a faster development experience. Since
Docker wouldn't do hot reloading for me, that meant that every change --
even one line changes -- would take almost a minute to actually run and
test. We also gain the ability to run test suites and linters a little
more easily, since we don't have to coordinate them with Docker.

A few things that are left to do: continue splitting up the
docker-compose file, find a way to reconnect the api service with
PostgreSQL and Redis, and find a way to connect Nginx to api and
frontend services.